+++ packages/dev/compiler/plugins/assets-module.ts
@@ -1,6 +1,8 @@
-import type { ResolvedWorkerConfig } from 'compiler/utils/config.js';
 import type { OnLoadResult, OnResolveArgs, OnResolveResult, Plugin, PluginBuild } from 'esbuild';
 import * as pkg from 'glob';
+import { normalize } from 'node:path';
+
+import type { ResolvedWorkerConfig } from '../utils/config.js';
 
 const { glob } = pkg;
 
@@ -10,7 +12,7 @@ const NAMESPACE = 'assets-module';
 export default function assetsPlugin(config: ResolvedWorkerConfig): Plugin {
   async function setup(build: PluginBuild) {
     const onResolve = ({ path }: OnResolveArgs): OnResolveResult => ({
-      path,
+      path: normalize(path),
       namespace: NAMESPACE,
       watchDirs: [config.assetsBuildDirectory],
     });